refactor(app): Update app component logic and routing configuration
- Refactored the app component to improve readability and maintainability.
- Updated the app component to use the vue-facing-decorator package for component and setup decorators.
- Replaced the deprecated useRouter function with the updated version from the vue-router package.
- Updated the app router configuration to use the HelloWorld component instead of the Home component.
- Removed unused files related to the app component.

This commit improves the structure and organization of the app component code, updates the routing configuration, and removes unnecessary files.
