--- src/pages/app/index.ts
@@ -1,3 +1,0 @@
-import AppView from './app.vue';
-
-export default AppView;
