--- src/pages/not-found/index.ts
@@ -1,2 +1,0 @@
-export { default } from './not-found.vue';
-
