--- src/pages/not-found/not-found.scss
@@ -1,3 +1,0 @@
-.not-found {
-    // Write your styles here.
-}
