--- src/pages/not-found/not-found.ts
@@ -1,42 +1,0 @@
-import { Component, Vue } from 'vue-property-decorator';
-
-@Component({
-  components: {},
-  name: 'not-found',
-})
-class NotFound extends Vue {
-  // --------------------------------------------------------------------------
-  // Fields
-  // --------------------------------------------------------------------------
-
-  // --------------------------------------------------------------------------
-  // Constructor
-  // --------------------------------------------------------------------------
-  constructor() {
-    super();
-  }
-
-  // --------------------------------------------------------------------------
-  // Accessors
-  // --------------------------------------------------------------------------
-
-  // --------------------------------------------------------------------------
-  // Methods
-  // --------------------------------------------------------------------------
-
-  // --------------------------------------------------------------------------
-  // Event Handlers
-  // --------------------------------------------------------------------------
-
-  // --------------------------------------------------------------------------
-  // Lifecycle Hooks
-  // --------------------------------------------------------------------------
-  public mounted() {
-    // TODO: stuff to do when this component loads.
-  }
-}
-
-export {
-  NotFound as default,
-  NotFound,
-};
