--- src/views/home-view/index.ts
@@ -1,1 +1,0 @@
-export { default } from './home-view.vue';
