+++ src/views/not-found/index.ts
@@ -1,0 +1,2 @@
+export { default } from './not-found.vue';
+
