+++ src/views/not-found/not-found.scss
@@ -1,0 +1,3 @@
+.not-found {
+    // Write your styles here.
+}
